Create_Prompts/requirements.txt
# cmake
pyarrow==17.0.0
torch
torchvision
transformers
# try transformers==4.51.1 if issues arise
accelerate
datasets
Pillow
huggingface_hub
tqdm
protobuf
sentencepiece 
safetensors